change type of timeoutMills from 'int' to 'long'

+++ streamis-jobmanager/streamis-job-manager/streamis-job-manager-base/src/main/java/com/webank/wedatasphere/streamis/jobmanager/manager/hook/StreamisJobShutdowHook.java
@@ -6,8 +6,8 @@ public interface StreamisJobShutdowHook {
 
     String getName();
 
-    void doBeforeJobShutdown(String jobId, String projectName, String jobName, int timeoutMills, Map<String, Object> params);
+    void doBeforeJobShutdown(String jobId, String projectName, String jobName, long timeoutMills, Map<String, Object> params);
 
-    void doAfterJobShutdown(String jobId, String projectName, String jobName, int timeoutMills, Map<String, Object> params);
+    void doAfterJobShutdown(String jobId, String projectName, String jobName, long timeoutMills, Map<String, Object> params);
 
 }
